code refactoring to reflect the inclusion of GameEntity class.
in order to keep the objects clean a new class -- GameEntity -- was
introduced. Entity was stripped down to contain only the things every
class deriving from it needs. all it has now is the id and the states.
probably i'll have to strip that further down, as the action class
doesn't need the states.