register Action and ActionLua Class.
previously the ActionLua Class was registered under the Action name.
however, this breaks the inheritance and makes it impossible to use
inherited methods. so the Action class was registered -- sans the
constructor, as the class is abstract --, and the ActionLua class is
registered under the ActionLua name.

ActionLua inherits from Action *and* from Entity. somehow slb can't cope
with abstract classes properly, so the Entity methods were not present
in ActionLua.