added a function for the run() method to determine whether to drop the action.
action was previously dropped from the execution queue the first time it
did what it does. now it is dropped under the condition that the done()
method returns true; what it does in its standard form. to define your
own condition to drop the action you have to reimplement the done()
method.