implemented fully supported EntityType behaviour.
EntityType objects are used as a fixed template for GameEntity objects.
they provide them with states, the value of which must not be changed,
as they may be assigned to more then just one entity, hence have to
deliver always the same value of this state.

if, however, the state of the type is going to be changed, a mechanism
	has been put in place to instantiate a new state object with the new
	value, and assigned directly to the entity. the next query for the
	state will return the state of the entity and not that of the type
	object anymore.