reworked the state handling.
now we get the packed version of a state, if the state comes from the
EntityType object. this newly created -- packed -- state is added
to the entity.

this state returns the value of the internal state pointer's value
unless the value gets overwritten. in this case the inter state pointer
is set to 0 and the new value is set directly to the state's value.