major refactoring of container classes to use smart pointers.
all container are now taking the ownership of their objects. in
particular

	* assigned child entities
	* assigned states
	* assigned actions
	* assigned form

are packed in a shared_ptr, which in turn is pushed into a container.

this results in all those objects getting deleted in the destructor
of their owner object.

the exception to this rule are the space objects. they don't take
the ownership of the assigned objects.