draw block considering the current row state.
current row is a state in the field's form. the row is computed
taking the current step and dividing it by the block height --
which is 40 px.

later it'd be possible to work with form's additional attributs
in order to define the speed of the block -- acceleration (which
is not yet in the drawable object).