core.graphics: rework the opengl backend to work with buffer objects.
use buffer objects to push the vertex data to the graphics first,
and render the buffer objects later all at once. the colour values
are pushed as first three float values within the position vertices
in order to be able to alter this data without the need for a new
buffer object being created and pushed ... in theory.