model: implemented model management.
in order to be able to handle memory in a better way, we have to be
in control as to how and where objects can be created. model management
is the first implementation of central place for creation of game
entities. no memory alignment was implemented in this commit though.