graphics: extend createWindow method.
it is now possible to create a fullscreen window. further more, no
window is created on init() of graphics module. hence the input
module has to be notified when the window is created -- as usually
the input module is initialised prior to creating a window. this
notification happens in the graphics module in the createWindow
method.